cut code from base.py to provider.py
